+++ app/src/main/java/com/example/weather/model/CityClimate.kt
@@ -1,0 +1,8 @@
+package com.example.weather.model
+
+data class CityClimate constructor(
+    val id: Int = 0,
+    val name: String = "",
+    val coord: Coord = Coord(),
+    val country: String = "",
+)
