+++ app/src/main/java/com/example/weather/model/ClimateByDay.kt
@@ -1,0 +1,9 @@
+package com.example.weather.model
+
+data class ClimateByDay(
+    val dt: Long,
+    val humidity: Double,
+    val pressure: Double,
+    val temp: Temp,
+    val wind_speed: Double
+)
