+++ app/src/main/java/com/example/weather/model/Clouds.kt
@@ -1,0 +1,5 @@
+package com.example.weather.model
+
+data class Clouds(
+    val all: Int
+)
