+++ app/src/main/java/com/example/weather/model/Coord.kt
@@ -1,0 +1,6 @@
+package com.example.weather.model
+
+data class Coord constructor(
+    val lon: Double = 0.0,
+    val lat: Double = 0.0
+)
