+++ app/src/main/java/com/example/weather/model/ForecastBy3h.kt
@@ -1,0 +1,12 @@
+package com.example.weather.model
+
+data class ForecastBy3h(
+    val dt: Long,
+    val main: Main,
+    val weather: List<Weather>,
+    val clouds: Clouds,
+    val wind: Wind,
+    val sys: Sys,
+    val dt_txt: String
+)
+
