+++ app/src/main/java/com/example/weather/model/Sys.kt
@@ -1,0 +1,5 @@
+package com.example.weather.model
+
+data class Sys(
+    val pod: String
+)
