+++ app/src/main/java/com/example/weather/model/Temp.kt
@@ -1,0 +1,9 @@
+package com.example.weather.model
+
+data class Temp(
+    val average: Double,
+    val average_max: Double,
+    val average_min: Double,
+    val record_max: Double,
+    val record_min: Double
+)
