+++ app/src/main/java/com/example/weather/model/Weather.kt
@@ -1,0 +1,8 @@
+package com.example.weather.model
+
+data class Weather(
+    val id: Int,
+    val main: String,
+    val description: String,
+    val icon: String
+)
