+++ app/src/main/java/com/example/weather/model/Wind.kt
@@ -1,0 +1,6 @@
+package com.example.weather.model
+
+data class Wind(
+    val speed: Double,
+    val deg: Double
+)
